Rename the none-response-type feature namespace to NoneResponse
Move Features/NoneResponseType to Features/NoneResponse (folder and namespace)
and point the consuming test host at it. Behaviour is unchanged — the none
response type support and its EnableNoneResponseType() opt-in keep their name
and public shape; only the containing namespace is shortened.
